fixed the pose double float bug(what the fuck)

+++ src/util.cpp
@@ -40,7 +40,7 @@ void video2images(const string &video_name,const string &image_dir){
  * @brief: x = K * X ==> K' * x
  * */
 cv::Point2f pixel2cam(const cv::Point2f &p, const cv::Mat & K){
-	return cv::Point2d(
+	return cv::Point2f(
 			(static_cast<float>(p.x) - K.at<float>(0,2)) / K.at<float>(0,0),
 			(static_cast<float>(p.y) - K.at<float>(1,2)) / K.at<float>(1,1)
 			);
